Updated query syntax for bibtex-ruby.
There seems to be a change in bibtex-ruby where bibtex-defined fields
such as public that I had are no longer easily indexable from the result
of the queries.  Instead, it has to be explicitly added in the query.

For example,

    ```bib.query('@*') { |b| (bib.public == 'yes') }``` no longer works.

It needs to be replaced with:

```bib.query('@*[public=yes]')```